code/requirements.txt
flask
gunicorn
prometheus-flask-exporter
kafka-python